+++ 05-ScenceManager/Koopas.h
@@ -19,6 +19,7 @@
 #define KOOPAS_ANI_SHELL		1
 #define KOOPAS_ANI_SHELLMOVING	2
 #define KOOPAS_ANI_FLYING		3
+
 #define TIME_RESTORE_MOVE 4000
 
 enum class KoopaState
@@ -35,7 +36,7 @@ class CKoopas : public CGameObject
 	virtual void GetBoundingBox(float& left, float& top, float& right, float& bottom);
 	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
 	virtual void Render();
-	DWORD startShellTime;
+	ULONGLONG startShellTime;
 	float velocity;
 public:
 	CKoopas();
