+++ 05-ScenceManager/RedKoopas.h
@@ -31,7 +31,7 @@ class RedKoopas : public CGameObject
 	virtual void GetBoundingBox(float& left, float& top, float& right, float& bottom);
 	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
 	virtual void Render();
-	DWORD startShellTime;
+	ULONGLONG startShellTime;
 	float velocity;
 public:
 	RedKoopas();
